# app/Http/Controllers/Api/TransactionController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Controller;
use App\Models\StockTransaction;
use App\Models\Item;
use App\Models\ActivityLog;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Validator;
use OpenApi\Attributes as OA;

class TransactionController extends Controller
{
    #[OA\Get(
        path: '/api/transactions',
        operationId: 'getTransactionList',
        summary: 'Ambil daftar transaksi stok',
        description: 'Mendapatkan seluruh riwayat transaksi mutasi stok (inbound & outbound) beserta detail barang, kategori, dan pengguna yang mencatat transaksi. Memerlukan hak akses Admin atau Staff.',
        security: [['bearerAuth' => []]],
        tags: ['Transactions'],
        responses: [
            new OA\Response(
                response: 200,
                description: 'Berhasil mengambil daftar transaksi',
                content: new OA\JsonContent(
                    properties: [
                        new OA\Property(property: 'success', type: 'boolean', example: true),
                        new OA\Property(
                            property: 'data',
                            type: 'array',
                            items: new OA\Items(
                                properties: [
                                    new OA\Property(property: 'id', type: 'integer', example: 1),
                                    new OA\Property(property: 'item_id', type: 'integer', example: 1),
                                    new OA\Property(property: 'user_id', type: 'integer', example: 1),
                                    new OA\Property(property: 'type', type: 'string', enum: ['in', 'out'], example: 'in'),
                                    new OA\Property(property: 'quantity', type: 'integer', example: 50),
                                    new OA\Property(property: 'notes', type: 'string', nullable: true, example: 'Restock GPU bulanan'),
                                    new OA\Property(property: 'status', type: 'string', enum: ['pending', 'completed'], example: 'completed'),
                                    new OA\Property(property: 'created_at', type: 'string', format: 'date-time'),
                                    new OA\Property(property: 'updated_at', type: 'string', format: 'date-time'),
                                    new OA\Property(
                                        property: 'item',
                                        type: 'object',
                                        properties: [
                                            new OA\Property(property: 'id', type: 'integer', example: 1),
                                            new OA\Property(property: 'name', type: 'string', example: 'RTX 4060 Ti'),
                                            new OA\Property(
                                                property: 'category',
                                                type: 'object',
                                                properties: [
                                                    new OA\Property(property: 'id', type: 'integer', example: 1),
                                                    new OA\Property(property: 'name', type: 'string', example: 'GPU')
                                                ]
                                            )
                                        ]
                                    ),
                                    new OA\Property(
                                        property: 'user',
                                        type: 'object',
                                        properties: [
                                            new OA\Property(property: 'id', type: 'integer', example: 1),
                                            new OA\Property(property: 'name', type: 'string', example: 'Admin')
                                        ]
                                    )
                                ]
                            )
                        )
                    ]
                )
            ),
            new OA\Response(response: 401, description: 'Unauthorized')
        ]
    )]
    public function index()
    {
        $transactions = StockTransaction::with(['item.category', 'user'])
            ->orderBy('created_at', 'desc')
            ->get();

        return response()->json([
            'success' => true,
            'data' => $transactions
        ]);
    }

    #[OA\Post(
        path: '/api/transactions',
        operationId: 'storeTransaction',
        summary: 'Buat transaksi stok baru',
        description: 'Mencatat transaksi mutasi stok baru (inbound/outbound). Jika outbound & status langsung completed, stok barang akan divalidasi dan dikurangi. Memerlukan hak akses Admin.',
        security: [['bearerAuth' => []]],
        tags: ['Transactions'],
        requestBody: new OA\RequestBody(
            required: true,
            content: new OA\JsonContent(
                required: ['item_id', 'type', 'quantity'],
                properties: [
                    new OA\Property(property: 'item_id', type: 'integer', example: 1),
                    new OA\Property(property: 'type', type: 'string', enum: ['in', 'out'], example: 'in'),
                    new OA\Property(property: 'quantity', type: 'integer', example: 10),
                    new OA\Property(property: 'notes', type: 'string', example: 'Pemasukan barang dari Gigabyte'),
                    new OA\Property(property: 'status', type: 'string', enum: ['pending', 'completed'], default: 'completed', example: 'completed')
                ]
            )
        ),
        responses: [
            new OA\Response(
                response: 201,
                description: 'Transaksi berhasil ditambahkan',
                content: new OA\JsonContent(
                    properties: [
                        new OA\Property(property: 'success', type: 'boolean', example: true),
                        new OA\Property(property: 'message', type: 'string', example: 'Transaksi berhasil ditambahkan'),
                        new OA\Property(property: 'data', type: 'object')
                    ]
                )
            ),
            new OA\Response(response: 400, description: 'Stok barang tidak mencukupi untuk outbound'),
            new OA\Response(response: 401, description: 'Unauthorized'),
            new OA\Response(response: 403, description: 'Forbidden - Hanya untuk Admin'),
            new OA\Response(response: 422, description: 'Validasi input gagal')
        ]
    )]
    public function store(Request $request)
    {
        $validator = Validator::make($request->all(), [
            'item_id' => 'required|exists:items,id',
            'type' => 'required|in:in,out',
            'quantity' => 'required|integer|min:1',
            'notes' => 'nullable|string',
            'status' => 'nullable|in:pending,completed',
        ]);

        if ($validator->fails()) {
            return response()->json([
                'success' => false,
                'message' => 'Validasi input gagal',
                'errors' => $validator->errors(),
            ], 422);
        }

        $item = Item::find($request->item_id);
        $type = $request->type;
        $quantity = intval($request->quantity);
        $status = $request->input('status', 'completed');
        $user = auth()->user();
        $userId = $user ? $user->id : 1;
        $userName = $user ? $user->name : 'Admin';

        // Jika transaksi outbound langsung status completed, cek stok mencukupi
        if ($type === 'out' && $status === 'completed' && $item->stock < $quantity) {
            return response()->json([
                'success' => false,
                'message' => 'Stok barang tidak mencukupi untuk melakukan transaksi outbound',
            ], 400);
        }

        // Simpan transaksi
        $tx = StockTransaction::create([
            'item_id' => $item->id,
            'user_id' => $userId,
            'type' => $type,
            'quantity' => $quantity,
            'notes' => $request->notes,
            'status' => $status,
        ]);

        // Sesuaikan stok jika statusnya langsung completed
        if ($status === 'completed') {
            if ($type === 'in') {
                $item->increment('stock', $quantity);
            } else {
                $item->decrement('stock', $quantity);
            }

            // Log ke audit trail ActivityLog
            $categoryTypes = [
                1 => 'gpu', 2 => 'cpu', 3 => 'ram', 4 => 'ssd',
                5 => 'mb', 6 => 'psu', 7 => 'case', 8 => 'cooling'
            ];
            $itemType = $categoryTypes[$item->category_id] ?? 'gpu';

            ActivityLog::create([
                'user_id' => $userId,
                'user_name' => $userName,
                'action' => $type === 'in' ? 'Restock' : 'Paid',
                'description' => "{$userName} membuat transaksi stok {$type} untuk {$item->name} sebanyak {$quantity} Unit",
                'item_type' => $itemType,
                'amount' => ($type === 'in' ? '+' : '-') . $quantity . ' Unit',
                'order_id' => '#' . str_pad($tx->id, 6, '0', STR_PAD_LEFT),
            ]);
        }

        return response()->json([
            'success' => true,
            'message' => 'Transaksi berhasil ditambahkan',
            'data' => $tx->load(['item', 'user']),
        ], 201);
    }

    #[OA\Put(
        path: '/api/transactions/{id}',
        operationId: 'updateTransactionStatus',
        summary: 'Update status transaksi (completed / pending)',
        description: 'Memperbarui status transaksi. Sistem akan otomatis menyesuaikan stok barang (increment/decrement/rollback) sesuai jenis transaksi dan status baru. Memerlukan hak akses Admin.',
        security: [['bearerAuth' => []]],
        tags: ['Transactions'],
        parameters: [
            new OA\Parameter(
                name: 'id',
                in: 'path',
                required: true,
                description: 'ID Transaksi',
                schema: new OA\Schema(type: 'integer', example: 1)
            )
        ],
        requestBody: new OA\RequestBody(
            required: true,
            content: new OA\JsonContent(
                required: ['status'],
                properties: [
                    new OA\Property(property: 'status', type: 'string', enum: ['pending', 'completed'], example: 'completed')
                ]
            )
        ),
        responses: [
            new OA\Response(
                response: 200,
                description: 'Status transaksi berhasil diperbarui',
                content: new OA\JsonContent(
                    properties: [
                        new OA\Property(property: 'success', type: 'boolean', example: true),
                        new OA\Property(property: 'message', type: 'string', example: 'Status transaksi berhasil diperbarui'),
                        new OA\Property(property: 'data', type: 'object')
                    ]
                )
            ),
            new OA\Response(response: 400, description: 'Penyesuaian stok gagal / stok tidak cukup'),
            new OA\Response(response: 401, description: 'Unauthorized'),
            new OA\Response(response: 403, description: 'Forbidden - Hanya untuk Admin'),
            new OA\Response(response: 404, description: 'Transaksi tidak ditemukan'),
            new OA\Response(response: 422, description: 'Validasi gagal')
        ]
    )]
    public function update(Request $request, string $id)
    {
        $tx = StockTransaction::find($id);

        if (!$tx) {
            return response()->json([
                'success' => false,
                'message' => 'Transaksi tidak ditemukan',
            ], 404);
        }

        $validator = Validator::make($request->all(), [
            'status' => 'required|in:pending,completed',
        ]);

        if ($validator->fails()) {
            return response()->json([
                'success' => false,
                'message' => 'Validasi gagal',
                'errors' => $validator->errors(),
            ], 422);
        }

        $oldStatus = $tx->status;
        $newStatus = $request->status;
        $item = Item::find($tx->item_id);
        $user = auth()->user();
        $userId = $user ? $user->id : 1;
        $userName = $user ? $user->name : 'Admin';

        if ($oldStatus !== $newStatus) {
            // Skenario 1: Pending -> Completed (Stok harus disesuaikan)
            if ($oldStatus === 'pending' && $newStatus === 'completed') {
                if ($tx->type === 'out' && $item->stock < $tx->quantity) {
                    return response()->json([
                        'success' => false,
                        'message' => 'Stok tidak mencukupi untuk menyelesaikan transaksi outbound ini',
                    ], 400);
                }

                if ($tx->type === 'in') {
                    $item->increment('stock', $tx->quantity);
                } else {
                    $item->decrement('stock', $tx->quantity);
                }
            }
            // Skenario 2: Completed -> Pending (Stok harus di-rollback)
            elseif ($oldStatus === 'completed' && $newStatus === 'pending') {
                // Rollback transaksi inbound (stok berkurang)
                if ($tx->type === 'in') {
                    if ($item->stock < $tx->quantity) {
                        return response()->json([
                            'success' => false,
                            'message' => 'Gagal mengubah status menjadi Pending. Stok barang saat ini kurang dari kuantitas transaksi inbound ini.',
                        ], 400);
                    }
                    $item->decrement('stock', $tx->quantity);
                } 
                // Rollback transaksi outbound (stok bertambah kembali)
                else {
                    $item->increment('stock', $tx->quantity);
                }
            }

            $tx->status = $newStatus;
            $tx->save();

            // Log ke audit trail ActivityLog
            $categoryTypes = [
                1 => 'gpu', 2 => 'cpu', 3 => 'ram', 4 => 'ssd',
                5 => 'mb', 6 => 'psu', 7 => 'case', 8 => 'cooling'
            ];
            $itemType = $categoryTypes[$item->category_id] ?? 'gpu';

            ActivityLog::create([
                'user_id' => $userId,
                'user_name' => $userName,
                'action' => 'Update',
                'description' => "{$userName} mengubah status transaksi #{$tx->id} ({$tx->type}) menjadi " . ucfirst($newStatus),
                'item_type' => $itemType,
                'amount' => 'Status',
                'order_id' => '#' . str_pad($tx->id, 6, '0', STR_PAD_LEFT),
            ]);
        }

        return response()->json([
            'success' => true,
            'message' => 'Status transaksi berhasil diperbarui',
            'data' => $tx->load(['item', 'user']),
        ], 200);
    }
}