# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management API Documentation",
        "description": "Dokumentasi API untuk Sistem Inventaris Tahap 1-4",
        "contact": {
            "email": "admin@inventaris.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server Utama"
        }
    ],
    "paths": {
...
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Ambil daftar transaksi stok",
                "description": "Mendapatkan seluruh riwayat transaksi mutasi stok (inbound & outbound) beserta detail barang, kategori, dan pengguna yang mencatat transaksi. Memerlukan hak akses Admin atau Staff.",
                "operationId": "getTransactionList",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar transaksi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "item_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "in",
                                                        "enum": [
                                                            "in",
                                                            "out"
                                                        ]
                                                    },
                                                    "quantity": {
                                                        "type": "integer",
                                                        "example": 50
                                                    },
                                                    "notes": {
                                                        "type": "string",
                                                        "example": "Restock GPU bulanan",
                                                        "nullable": true
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "completed",
                                                        "enum": [
                                                            "pending",
                                                            "completed"
                                                        ]
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "item": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "RTX 4060 Ti"
                                                            },
                                                            "category": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "GPU"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Admin"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Buat transaksi stok baru",
                "description": "Mencatat transaksi mutasi stok baru (inbound/outbound). Jika outbound & status langsung completed, stok barang akan divalidasi dan dikurangi. Memerlukan hak akses Admin.",
                "operationId": "storeTransaction",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "item_id",
                                    "type",
                                    "quantity"
                                ],
                                "properties": {
                                    "item_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "in",
                                        "enum": [
                                            "in",
                                            "out"
                                        ]
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Pemasukan barang dari Gigabyte"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "completed",
                                        "default": "completed",
                                        "enum": [
                                            "pending",
                                            "completed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaksi berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transaksi berhasil ditambahkan"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Stok barang tidak mencukupi untuk outbound"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Hanya untuk Admin"
                    },
                    "422": {
                        "description": "Validasi input gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions/{id}": {
            "put": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Update status transaksi (completed / pending)",
                "description": "Memperbarui status transaksi. Sistem akan otomatis menyesuaikan stok barang (increment/decrement/rollback) sesuai jenis transaksi dan status baru. Memerlukan hak akses Admin.",
                "operationId": "updateTransactionStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Transaksi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "completed",
                                        "enum": [
                                            "pending",
                                            "completed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status transaksi berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Status transaksi berhasil diperbarui"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Penyesuaian stok gagal / stok tidak cukup"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Hanya untuk Admin"
                    },
                    "404": {
                        "description": "Transaksi tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}